requirements.txt
opencv-python
numpy
keyboard
Pillow